Refine physical address mapping of flash chips.
Previously we did map the physical memory corresponding to all chips,
even for SPI chips where it is completely unnecessary because we do
not use it in any SPI-capable programmer.
Also, the mapping was never undone before the program exited.
Mapping of additional flash registers was done within probing methods
instead of centrally.
Besides all of that, this patch also adds proper error handling for
mapping-related operations.